+++ bank/src/main/java/hei/shool/bank/dtos/requests/TransferRequest.java
@@ -1,12 +1,10 @@
 package hei.shool.bank.dtos.requests;
 
-import java.math.BigDecimal;
-
 public record TransferRequest (
         String senderAccountNumber,
         String receiverAccountNumber,
 
-        BigDecimal amount,
+        Double amount,
 
         String description
 ){}
